loggerd: split loggerd.cc into three files
