+++ selfdrive/loggerd/SConscript
@@ -7,7 +7,7 @@ libs = [logger_lib, common, cereal, messaging, visionipc,
         'avformat', 'avcodec', 'swscale', 'avutil',
         'yuv', 'bz2', 'OpenCL']
 
-src = ['loggerd.cc']
+src = ['main.cc', 'loggerd.cc']
 if arch in ["aarch64", "larch64"]:
   src += ['omx_encoder.cc']
   libs += ['OmxCore', 'gsl', 'CB'] + gpucommon
@@ -24,7 +24,7 @@ if arch == "Darwin":
   del libs[libs.index('OpenCL')]
   env['FRAMEWORKS'] = ['OpenCL']
 
-env.Program(src, LIBS=libs)
+env.Program('loggerd', src, LIBS=libs)
 env.Program('bootlog.cc', LIBS=libs)
 
 if GetOption('test'):
